arm64: dts: rockchip: only use one opp table for rk3368
The rk3368 does have 2 cpu clusters but both clusters are supplied by the
same regulator. This seems to stem from the notion that only one of the
clusters of 4 cores should be active (little or big). Both clusters seem
to be able to reach 1.2GHz while the big cluster also should be able
to reach 1.5GHz.

The big-little switcher seems deprecated and in general it seems more
desirable for people I talked to, to have all 8 cores under kernel control
and usable concurrently.

To be able to switch cluster-frequencies independently it would be
necessary to let the slower cluster run at the voltage the faster cluster
uses, but the current cpufreq and regulator kernel frameworks cannot cope
with this and having one cluster run slow on a higher than necessary
voltage is also wasteful and while experimenting I even noticed unexpected
usb disconnects even though the frequency-setting failed as expected.

So again from people I talked to it is more desirable to have all cores
reach their shared maximum frequency (1.2GHz in the future) and running
at the lowest possible voltage and thus we drop the second opp table and
let both cluster share the same table.

Fixes: 6f2dea1f5fdb ("arm64: dts: rockchip: Add basic cpu frequencies for RK3368")
Signed-off-by: Heiko Stuebner <heiko@sntech.de>