drm/lima: add basic support for Mali-450
Mali-450 is a bit different compared to Mali-400.
Not only does it support up to 8 PPs grouped into 2 groups but also
uses separate L2 caches for GP, PP0-3 and PP4-7.

Additionally it also provides some new components like a broadcast
unit, a load balancer and dma block.

This only supports the Mali-450 in a trimmed down variant using the
separate caches but limiting the PPs to the first group and thus a
maximum of 4 PPs and ignoring the other blocks for now.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>